materials/day4_quarto/pyproject.toml
# pyproject.toml
[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[project]
name            = "breast-cancer-classifier"
version         = "0.1.0"
description     = "Clasificador de cáncer de mama — Workshop ML Día 2"
requires-python = ">=3.10"

dependencies = [
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "PyYAML>=6.0",
    "wandb>=0.16.0",
    "dvc>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length    = 99
target-version = "py310"
select         = ["E", "W", "F", "I", "N", "UP"]
ignore         = ["E501"]
exclude        = [".venv", "__pycache__", "outputs", "data"]

[tool.ruff.format]
quote-style  = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths    = ["tests"]
python_files = ["test_*.py"]
addopts      = "-v --tb=short"
